Meeting report: refactor (WIP)
report display refactor
