+++ lib/frontend/src/actions/meetings.js
@@ -20,6 +20,11 @@ export const toggleEdition = value => ({
   value: value,
 })
 
+export const updateLastReportDate = value => ({
+  type: 'UPDATE_LAST_REPORT',
+  value: value,
+})
+
 export const updateMeeting = (data, dataType) => ({
   type: 'MEETING_UPDATE',
   data,
@@ -95,118 +100,16 @@ export const updateReportsList = (event, history = null) => dispatch => {
   }
 }
 
-const sortPresents = presents => {
-  const presentsList = []
-  for (let i = 0; i < presents.length; i++) {
-    presentsList.push({ name: presents[i].name, checked: presents[i].checked })
-  }
-  presentsList.sort((a, b) => b.checked - a.checked)
-  return presentsList
-}
-
-export const submitReport = (event, meetingType, history) => async (
-  dispatch,
-  getState
-) => {
-  event.persist()
-
+export const submitReport = history => async (dispatch, getState) => {
   let response
   let json = {}
   const state = getState()
-  let indicators = ''
-  let actions = ''
-  let milestones = ''
-  let issues = ''
-  let users = ''
-  let fullcontent = ''
-
-  for (let i = 0; i < event.target.form.length; i++) {
-    switch (event.target.form[i].id) {
-      case 'indicateurs':
-        indicators += `
-* ${event.target.form[i].name} : ${
-          event.target.form[i].value === '' ? 0 : event.target.form[i].value
-        }
-         `
-        break
-      case 'actions':
-        actions += `
-* [${event.target.form[i].checked ? 'x' : ' '}] ${event.target.form[i].name}
-        `
-        break
-      case 'milestones':
-        milestones += `
-* ${event.target.form[i].name}${
-          event.target.form[i].value === ''
-            ? ''
-            : `: ${event.target.form[i].value}`
-        }
-        `
-        break
-      case 'agenda':
-        issues += `
-  * ${event.target.form[i].name}${
-          event.target.form[i].value === ''
-            ? ''
-            : `: ${event.target.form[i].value}`
-        }
-          `
-        break
-    }
-  }
-
-  const sortedUsers = sortPresents(event.target.form.users)
-  for (let i = 0; i < sortedUsers.length; i++) {
-    users += `
-* [${sortedUsers[i].checked ? 'x' : ' '}] ${sortedUsers[i].name}
-    `
-  }
-
-  fullcontent = `
-
-### Presents:
-
-${users}
-
-`
-
-  if (meetingType === 'Triage') {
-    fullcontent += `
-
-  ### Recurrent Actions:
-
-  ${actions}
-
-  ### Indicators:
-
-  ${indicators}
-
-  ### Milestones:
-
-  ${milestones}
-
-`
-  }
-
-  fullcontent += `
-
-  ### Ordre du jour:
-
-  #### Tickets
-
-  ${issues}
-
-  #### Autres
-
-  ${event.target.form.body.value}
-
-  `
 
   const newReport = {
-    circle_id: +event.target.form.circles.value,
-    report_type: event.target.form.meetingType.value,
+    circle_id: state.params.circle_id,
+    report_type: state.params.meeting_name,
     author_id: state.user.id,
-    content: fullcontent,
+    content: JSON.stringify(state.meeting.results),
   }
   try {
     response = await fetch('/api/reports', {
@@ -220,7 +123,7 @@ ${users}
   } catch (e) {
     return dispatch(
       setError('Erreur pendant la création du rapport'),
-      'meetings'
+      'meeting'
     )
   }
   if (response.status >= 200 && response.status < 300) {
@@ -228,23 +131,21 @@ ${users}
       json = await response.json()
       return history.push(`/meeting?report_id=${json.objects[0].report_id}`)
     } catch (e) {
-      return dispatch(
-        setError('La réponse ne contient pas de json', 'meetings')
-      )
+      return dispatch(setError('La réponse ne contient pas de json', 'meeting'))
     }
   } else {
     const jsonMessage = typeof json === 'undefined' ? '' : json.message
     return dispatch(
       setError(
         `${response.status}: ${response.statusText} ${jsonMessage}`,
-        'meetings'
+        'meeting'
       )
     )
   }
 }
 
 export const fetchReport = (search = null) => async (dispatch, getState) => {
-  let response
+  let response, content, meeting
   let json = {}
   const state = getState()
   const report = search ? search : reportIdFromState(state)
@@ -262,7 +163,6 @@ export const fetchReport = (search = null) => async (dispatch, getState) => {
   if (response.status >= 200 && response.status < 300) {
     try {
       json = await response.json()
-      dispatch(setResults(json.objects[0], 'meeting'))
     } catch (e) {
       console.error(e)
       return dispatch(setError('La réponse ne contient pas de json', 'meeting'))
@@ -275,6 +175,24 @@ export const fetchReport = (search = null) => async (dispatch, getState) => {
         'meeting'
       )
     )
+  }
+  if (json.objects[0]) {
+    try {
+      content = JSON.parse(json.objects[0].content)
+    } catch (e) {
+      // old version of meeting report (text only)
+      content = {
+        attendees: [],
+        actions: [],
+        indicators: [],
+        projects: [],
+        agenda: [],
+        content: json.objects[0].content,
+      }
+    }
+    delete json.objects[0].content
+    meeting = Object.assign({}, json.objects[0], content)
+    dispatch(setResults(meeting, 'meeting'))
   }
 }
 
@@ -436,7 +354,12 @@ export const fetchCircleAgendaIssues = (
 }
 
 export const getLastReport = params => async dispatch => {
-  if (params.circle_id === '' || params.meeting_name === '') {
+  if (
+    !params.circle_id ||
+    params.circle_id === '' ||
+    !params.meeting_name ||
+    params.meeting_name === ''
+  ) {
     return
   }
   let json, response
@@ -460,9 +383,8 @@ export const getLastReport = params => async dispatch => {
     try {
       json = await response.json()
       return dispatch(
-        updateMeeting(
-          json.objects[0] ? json.objects[0].created_at : null,
-          'lastReportDate'
+        updateLastReportDate(
+          json.objects[0] ? json.objects[0].created_at : null
         )
       )
     } catch (e) {
@@ -500,7 +422,7 @@ export const fetchMeetingData = params => async (dispatch, getState) => {
   promises.push(
     dispatch(getUsersListFromRoles(circle.roles, state.users.results))
   )
-
+  promises.push(dispatch(toggleEdition(false)))
   if (meetingType === 'Triage') {
     promises.push(dispatch(fetchCircleMilestonesAndIssues(circle)))
   }
