+++ lib/frontend/src/actions/milestones.js
@@ -116,7 +116,7 @@ export const fetchCircleMilestonesAndIssues = circle => async (
   }
   const state = getState()
   // eslint-disable-next-line prefer-destructuring
-  const lastReportDate = state.meeting.results.lastReportDate
+  const lastReportDate = state.meeting.lastReportDate
   const promises = []
   for (let i = 0; i < circle.circle_milestones.length; i++) {
     let { circleMilestone } = circle.circle_milestones[i]
