--- lib/frontend/src/components/Meeting.sass
@@ -1,30 +1,0 @@
-@import ../vars
-
-.Meeting
-  flex: 1
-  overflow: auto
-  position: relative
-
-  &__meeting
-    margin-left: 1em
-    margin-right: .5em
-
-  &__content
-    background: #e0e9f0
-    border: none
-    box-sizing: border-box
-    margin: 2em
-    padding: 1em
-    outline: none
-    width: 50em
-
-  &__date
-    color: #50595f
-    margin-left: .75em
-
-  &__editor
-    margin-top: 1em
-    width: 49em
-
-.editMeeting
-  margin-left: .5em
