+++ lib/frontend/src/components/MeetingReport/MeetingReport.sass
@@ -1,6 +1,6 @@
 @import ../vars
 
-.MeetingsReportCreation
+.MeetingReport
   flex: 1
   overflow: auto
   position: relative
@@ -28,6 +28,14 @@
     margin-top: 1em
     width: 49em
 
+  &__date
+    color: #50595f
+    margin-left: .75em
+
+  &__editor
+    margin-top: 1em
+    width: 49em
+
   &__due-date
     font-size: .8em
     margin-left: .5em
@@ -39,6 +47,14 @@
     color: #323232
     font-size: .8em
     margin-left: .5em
+
+  &__markdownContent
+    background: #e0e9f0
+    border: none
+    box-sizing: border-box
+    padding: .5em
+    outline: none
+    width: 44.5em
 
   &__meetings
     margin-left: 1em
