+++ lib/frontend/src/components/MeetingReportCreation.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+import MeetingReport from './MeetingReport'
+
+export default function MeetingReportCreation() {
+  return <MeetingReport isCreation />
+}
