+++ lib/frontend/src/reducer/index.js
@@ -291,7 +291,7 @@ const meeting = (state = initial.meeting, action) => {
     case 'MEETING_ON_EDITION':
       return {
         ...state,
-        is_in_edition: action.value,
+        isEditionDisabled: action.value,
       }
     case 'MEETING_UPDATE':
       return {
@@ -386,6 +386,11 @@ const meeting = (state = initial.meeting, action) => {
             return milestone
           }),
         },
+      }
+    case 'UPDATE_LAST_REPORT':
+      return {
+        ...state,
+        lastReportDate: action.value,
       }
     default:
       return handleResultsLoadingAndError(state, 'meeting', action)
