+++ lib/frontend/src/reducer/initial.js
@@ -63,9 +63,9 @@ export default {
       projects: [],
       agenda: [],
       content: '',
-      lastReportDate: null,
     },
-    is_in_edition: false,
+    isEditionDisabled: true,
+    lastReportDate: null,
   },
   meetings: {
     ...emptyResults,
